refactor(library-config): split otel process context reader
